Fixed gradient calculation
Issue with using the gradient has been resolved. The positions for the initial state of the time intergration were not consistent with the positions of the gradient parameters. As well asys was not properly calculating the time integration of the gradient dH due to an indexing error on the initial state.